pyproject.toml
[project]
name = "ziip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "deepl>=1.28.0",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "httpx>=0.28.1",
    "konlpy>=0.6.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "coverage>=7.13.1",
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
